Move test case to different module
