+++ test/service/test_repository_files.py
@@ -60,17 +60,8 @@ from azul.plugins.metadata.hca import (
 from azul.service.repository_service import (
     RepositoryService,
 )
-from azul.service.source_service import (
-    NotFound,
-    SourceService,
-)
 from azul.terra import (
-    TDRClient,
-    TDRSourceSpec,
     TerraClient,
-)
-from azul.types import (
-    JSON,
 )
 from azul_test_case import (
     DCP1TestCase,
@@ -178,72 +169,6 @@ class TestRepositoryFilesWithTDR(DCP2TestCase, RepositoryFilesTestCase):
                               return_value=file):
                 response = client.request('GET', str(azul_url), redirect=False)
             self.assertEqual(response.status, 404)
-
-
-class TestListSources(DCP2TestCase, LocalAppTestCase):
-
-    @classmethod
-    def app_name(cls) -> str:
-        return 'service'
-
-    mock_source_names = ['mock_snapshot_1', 'mock_snapshot_2']
-    make_mock_source_spec = 'tdr:bigquery:gcp:mock:{}'.format
-
-    @classmethod
-    def _sources(cls):
-        return {
-            cls.make_mock_source_spec(n): {'mirror': True}
-            for n in cls.mock_source_names
-        }
-
-    @patch.object(SourceService, '_get')
-    @patch.object(TDRClient, 'snapshot_names_by_id')
-    @patch.object(TDRClient, 'validate', new=MagicMock())
-    def test(self, mock_tdr_client__snapshot_names_by_id, mock_source_service__get):
-        # Includes extra sources to check that the endpoint only returns results
-        # for the current catalog
-        extra_sources = ['foo', 'bar']
-        mock_source_names_by_id = {
-            str(i): source_name
-            for i, source_name in enumerate(self.mock_source_names + extra_sources)
-        }
-        mock_tdr_client__snapshot_names_by_id.return_value = mock_source_names_by_id
-        client = http_client(log)
-        azul_url = furl(url=self.base_url,
-                        path='/repository/sources',
-                        query_params=dict(catalog=self.catalog))
-
-        def _list_sources(headers) -> JSON:
-            response = client.request('GET',
-                                      str(azul_url),
-                                      headers=headers)
-            self.assertEqual(response.status, 200)
-            return json.loads(response.data)
-
-        def _test(*, authenticate: bool, cache: bool):
-            with self.subTest(authenticate=authenticate, cache=cache):
-                response = _list_sources({'Authorization': 'Bearer foo_token'}
-                                         if authenticate else {})
-                self.assertEqual(response, {
-                    'sources': [
-                        {
-                            'sourceId': id,
-                            'sourceSpec': str(TDRSourceSpec.parse(self.make_mock_source_spec(name)))
-                        }
-                        for id, name in mock_source_names_by_id.items()
-                        if name not in extra_sources
-                    ]
-                })
-
-        mock_source_service__get.return_value = list(mock_source_names_by_id.keys())
-        _test(authenticate=True, cache=True)
-        _test(authenticate=False, cache=True)
-        mock_source_service__get.return_value = None
-        mock_source_service__get.side_effect = NotFound('foo_token')
-        with patch('azul.terra.TDRClient.snapshot_ids',
-                   return_value=mock_source_names_by_id.keys() | {'not_indexed'}):
-            _test(authenticate=True, cache=False)
-            _test(authenticate=False, cache=False)
 
 
 class TestRepositoryFilesWithDSS(DCP1TestCase,
